+++ microservices/object_detection/google_coral/config.py
@@ -1,4 +1,4 @@
-from detector_base import load_labels
+import detector_base
 
 # input shape
 W = 640
@@ -8,6 +8,7 @@ C = 3
 SAMPLE_XML_ANNOTATION_PATH = "annotation.xml"
 FOLDER_NAME = "ai4coral2021"
 PATH_TO_LABELS = 'models/coco_labels.txt'
-LABELS = load_labels(PATH_TO_LABELS)
+# PATH_TO_LABELS = 'models/demo_labels.txt'
 PATH_TO_MODEL = 'models/ssd_mobilenet_v2_coco_quant_postprocess.tflite'
+# PATH_TO_MODEL = 'models/core.tflite'
 THRESHOLD = 0.4
