--- microservices/object_detection/google_coral/object_detection_abc.py
@@ -1,10 +1,0 @@
-import abc
-
-class ObjectDetectionABC(abc.ABC):
-    @abc.abstractmethod
-    def get_detection(self, image_frame):
-        """ 
-            Return a Pascal VOC format annotation containig 
-            the detections made by the model on the image_frame
-        """
-        pass
