--- microservices/object_detection/google_coral/run.sh
@@ -1,5 +1,0 @@
-python3 detector_base.py \
---model models/ssd_mobilenet_v2_coco_quant_postprocess.tflite \
---labels models/coco_labels.txt \
---input images/image1.jpg \
---output images/image1_processed.jpg
