+++ microservices/object_detection/google_coral/run_trained.sh
@@ -1,0 +1,5 @@
+python3 detector_base.py \
+--model models/core.tflite \
+--labels models/demo_labels.txt \
+--input images/fish.jpg \
+--output images/fish_processed.jpg
